Move protobuf NodeId to pb crate under dev/restate/common
This commit moves the NodeId to the pb crate under dev/restate/common.
