+++ crates/node-ctrl-proto/src/lib.rs
@@ -8,10 +8,13 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+use restate_pb::restate::common;
+
 pub mod node_ctrl {
     #![allow(warnings)]
     #![allow(clippy::all)]
     #![allow(unknown_lints)]
+
     tonic::include_proto!("dev.restate.node_ctrl");
 
     pub const FILE_DESCRIPTOR_SET: &[u8] =
@@ -27,19 +30,4 @@ pub mod cluster_controller {
 
     pub const FILE_DESCRIPTOR_SET: &[u8] =
         tonic::include_file_descriptor_set!("cluster_controller_descriptor");
-
-    impl From<NodeId> for restate_types::NodeId {
-        fn from(node_id: NodeId) -> Self {
-            restate_types::NodeId::new(node_id.id, node_id.generation)
-        }
-    }
-
-    impl From<restate_types::NodeId> for NodeId {
-        fn from(node_id: restate_types::NodeId) -> Self {
-            NodeId {
-                id: node_id.id().into(),
-                generation: node_id.as_generational().map(|g| g.generation()),
-            }
-        }
-    }
 }
