+++ crates/pb/Cargo.toml
@@ -10,6 +10,7 @@ publish = false
 [features]
 default = []
 builtin-service = ["restate-types"]
+common = ["restate-types"]
 mocks = []
 
 [dependencies]
